[CHORE] Reorganization of the project structure
