+++ src/gestion_satisfactory/scraper/utils.py
