+++ src/gestion_satisfactory/utils/config/config.py
@@ -1,0 +1,72 @@
+raw_materials_1_ = {
+    'Limestone':1.,
+    'Iron Ore':1.,
+    'Copper Ore':1.,
+    'Coal':2.,
+    'Sulfur':2.,
+    'Raw Quartz':4.,
+    'Bauxite':8.,
+    'Caterium Ore':6.,
+    'Water':0.,
+    'Crude Oil':.5,
+    'Nitrogen Gas':6.,
+    'Uranium':12.,
+    'SAM':50.,
+    'Uranium Waste':10.,
+    'Plutonium Waste':10.,
+}
+
+raw_materials_2_ = {
+    'Mycelia':1.,
+    'Leaves':1.,
+    'Wood':1.,
+    'Blue Power Slug':1.,
+    'Purple Power Slug':1.,
+    'Yellow Power Slug':1.,
+    'Stinger Remains':1.,
+    'Hog Remains':1.,
+    'Hatcher Remains':1.,
+    'Spitter Remains':1.,
+    'FICSMAS Tree Branch':1.,
+    'FICSMAS Bow':1.,
+    'Actual Snow':1.,
+    'Candy Cane':1.,
+}
+
+raw_materials_limit_1_ = {
+    'Limestone':5000.,
+    'Iron Ore':5000.,
+    'Copper Ore':5000.,
+    'Coal':5000.,
+    'Sulfur':5000.,
+    'Raw Quartz':4000.,
+    'Bauxite':2000.,
+    'Caterium Ore':1000.,
+    'Water':10000.,
+    'Crude Oil':8000.,
+    'Nitrogen Gas':1000.,
+    'Uranium':1000.,
+    'SAM':500.,
+    'Uranium Waste':10.,
+    'Plutonium Waste':10.,
+}
+
+raw_materials_limit_2_ = {
+    'Mycelia':200.,
+    'Leaves':200.,
+    'Wood':200.,
+    'Blue Power Slug':200.,
+    'Purple Power Slug':200.,
+    'Yellow Power Slug':200.,
+    'Stinger Remains':200.,
+    'Hog Remains':200.,
+    'Hatcher Remains':200.,
+    'Spitter Remains':200.,
+    'FICSMAS Tree Branch':200.,
+    'FICSMAS Bow':200.,
+    'Actual Snow':200.,
+    'Candy Cane':200.,
+}
+
+df_factory_planner_columns = ["area", "factory", "line", "building_name", "nb_building", "power_usage", "rate/overclock", "recipe_name",
+    "duration", "min_consumption","max_consumption","ingredient_1","ingredient_amount_1","ingredient_2","ingredient_amount_2","ingredient_3","ingredient_amount_3","ingredient_4","ingredient_amount_4","product_1","product_amount_1","product_2","product_amount_2"]
