+++ src/gestion_satisfactory/utils/database/load_df.py
@@ -1,7 +1,7 @@
 import base64
 import pandas as pd
 
-from .connect_bdd import load_df
+from gestion_satisfactory.utils.database.connect_bdd import load_df
 
 
 def ReadPictureFile(wch_fl):
