+++ src/gestion_satisfactory/utils/display/grids.py
