+++ src/gestion_satisfactory/utils/optimization/optimize.py
